Extract interface from `ScrollResult` class.
