+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/Elasticsearch6Module.java
@@ -1,5 +1,6 @@
 package org.graylog.storage.elasticsearch6;
 
+import com.google.inject.assistedinject.FactoryModuleBuilder;
 import org.graylog.events.indices.EventIndexerAdapter;
 import org.graylog.events.search.MoreSearchAdapter;
 import org.graylog.storage.elasticsearch6.migrations.V20170607164210_MigrateReopenedIndicesToAliasesClusterStateES6;
@@ -9,6 +10,7 @@ import org.graylog2.indexer.cluster.NodeAdapter;
 import org.graylog2.indexer.fieldtypes.IndexFieldTypePollerAdapter;
 import org.graylog2.indexer.indices.IndicesAdapter;
 import org.graylog2.indexer.messages.MessagesAdapter;
+import org.graylog2.indexer.results.ScrollResult;
 import org.graylog2.indexer.searches.SearchesAdapter;
 import org.graylog2.migrations.V20170607164210_MigrateReopenedIndicesToAliases;
 import org.graylog2.plugin.PluginModule;
@@ -26,5 +28,8 @@ public class Elasticsearch6Module extends PluginModule {
         bind(IndexFieldTypePollerAdapter.class).to(IndexFieldTypePollerAdapterES6.class);
         bind(IndexToolsAdapter.class).to(IndexToolsAdapterES6.class);
         bind(V20170607164210_MigrateReopenedIndicesToAliases.ClusterState.class).to(V20170607164210_MigrateReopenedIndicesToAliasesClusterStateES6.class);
+
+        install(new FactoryModuleBuilder().build(ScrollResultES6.Factory.class));
+        bind(ScrollResult.Factory.class).to(ScrollResultES6.Factory.class);
     }
 }
