+++ graylog-storage-elasticsearch6/src/test/resources/org/graylog/storage/elasticsearch6/ScrollResultIT.json
